omnidet/requirements.txt
onnxmltools==1.9.0
onnxruntime==1.8.1
colorama
tqdm==4.62.1
tifffile==2021.8.8
fastremap==1.12.2
pytorch_model_summary==0.1.1
torch_dct==0.1.5
matplotlib==3.1.3
scikit_image==0.16.2
torch==1.9.0
mpmath==1.2.1
torchvision==0.10.0
opencv_python==4.5.3.56
pandas==1.3.2
numpy==1.21.2
Pillow==7.2.0
PyYAML==5.4.1
future
Shapely==1.7.1
tensorboard==2.6.0